Feature Update on GUI and chargeToVolt & chargeToSOC Scripts
Updated the arguments sent between GUI and Scripts and how they're used in the BatteryManager, DeviceConfig GUI files and the chargeToVolt and chargeToSOC files in preparation for testing.  Updated how the Triggers work in the script improving on the number of triggers possible and their implementation.
Also updated the device initialization script to reflect some other recent changes and updated the temperature sensors to receive readings from any channel regardless or order or usage.