Fixing error due to the inability to find eloadArgs
While Checking for the args of each device in script_initializeDevices, the program threw an errorCode of BAD_DEV_ARG thinking that the individual arguments were bad instead of just the group of arguments unavailable since they were not being used. Corrected the if statements to first check if args are available, if not skip device initialization. If args is not empty, and the arg needed is not available only then should error be caught.