Tested and fixed some bugs related to chargeToVolt from the GUI
Current implementation is able to start a test using a parallel pool, and send data as they are collected back to the GUI.
The main lingering bug is the parfeval code quitting before the the underlying test is complete due to errors in the timer function. Need a way to send commands to the parallel pool to take control back or something related to that.