Adding functionality for sending commands to parallel function evals
Added ability to stop, pause, and unpause the program. The additions consist of a file "script_checkGUICmd" to check for the commands and change the "testStatus" variable to either stop the running test or continue it. Continuing it after a pause duration also increases the trigger times if they are available.