Added a curr model containing balance current transformation matrix
Added the ability to transform the balancing currents to better emulate what actually happens in the balancer. Before, the primary cell currents were added to the PSUcurr, but because the some amount of balance current is shared with the stack and the efficiency is not 100%, this transformation matrix helps draw out the net cell current instead.