Initial Commit of the SubSampledModel method.
Change a few things to allow for model subsampling. This is going to allow the model to be aware of the inability of the balancer to set a variable current. It essentially now allows the MPC to work on estimating future states and control actions based on the timer control of the balancer.