Initial Commit of StructDataFormat
Data collection in this branch jas been sucessfully converted from a timeseries datatype to a structure to allow for modularity and to make it more intuitive.

The architecure for multi-battery testing has also been simplified. The previous method had an ID for each cell in a pack and an ID for the pack, but now, regardless of what configuration the pack is, all the user needs to provide is a "battID" and the information about the cell arrangement, and individual cell information. The cells are accessed by indices now not IDs.